+++ src/main/java/com/jamjam/bookjeok/domains/cart/controller/CartController.java
@@ -1,11 +1,11 @@
-package com.jamjam.bookjeok.domains.order.controller.cart;
+package com.jamjam.bookjeok.domains.cart.controller;
 
 import com.jamjam.bookjeok.common.dto.ApiResponse;
-import com.jamjam.bookjeok.domains.order.dto.cart.request.CartMemberIdRequest;
-import com.jamjam.bookjeok.domains.order.dto.cart.response.CartBookListResponse;
-import com.jamjam.bookjeok.domains.order.dto.cart.response.CartResponse;
-import com.jamjam.bookjeok.domains.order.dto.cart.request.CartRequest;
-import com.jamjam.bookjeok.domains.order.service.cart.CartService;
+import com.jamjam.bookjeok.domains.cart.dto.request.CartMemberIdRequest;
+import com.jamjam.bookjeok.domains.cart.dto.response.CartBookListResponse;
+import com.jamjam.bookjeok.domains.cart.dto.response.CartResponse;
+import com.jamjam.bookjeok.domains.cart.dto.request.CartRequest;
+import com.jamjam.bookjeok.domains.cart.service.CartService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
