+++ src/main/java/com/jamjam/bookjeok/domains/cart/dto/request/CartMemberIdRequest.java
@@ -1,4 +1,4 @@
-package com.jamjam.bookjeok.domains.order.dto.cart.request;
+package com.jamjam.bookjeok.domains.cart.dto.request;
 
 import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
