+++ src/main/java/com/jamjam/bookjeok/domains/cart/dto/request/CartRequest.java
@@ -1,4 +1,4 @@
-package com.jamjam.bookjeok.domains.order.dto.cart.request;
+package com.jamjam.bookjeok.domains.cart.dto.request;
 
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
