+++ src/main/java/com/jamjam/bookjeok/domains/cart/repository/CartRepository.java
@@ -1,6 +1,6 @@
-package com.jamjam.bookjeok.domains.order.repository.cart;
+package com.jamjam.bookjeok.domains.cart.repository;
 
-import com.jamjam.bookjeok.domains.order.entity.Cart;
+import com.jamjam.bookjeok.domains.cart.entity.Cart;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CartRepository extends JpaRepository<Cart, Long> {
