+++ src/test/java/com/jamjam/bookjeok/domains/cart/service/CartServiceImplTest.java
@@ -1,9 +1,9 @@
-package com.jamjam.bookjeok.domains.order.service.cart;
+package com.jamjam.bookjeok.domains.cart.service;
 
-import com.jamjam.bookjeok.domains.order.dto.cart.response.CartBookListResponse;
-import com.jamjam.bookjeok.domains.order.dto.cart.response.CartResponse;
-import com.jamjam.bookjeok.domains.order.dto.cart.request.CartRequest;
-import com.jamjam.bookjeok.exception.order.cart.CartBookNotFoundException;
+import com.jamjam.bookjeok.domains.cart.dto.response.CartBookListResponse;
+import com.jamjam.bookjeok.domains.cart.dto.response.CartResponse;
+import com.jamjam.bookjeok.domains.cart.dto.request.CartRequest;
+import com.jamjam.bookjeok.exception.cart.CartBookNotFoundException;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
